+++ packages/host-kit/package.json
@@ -19,6 +19,14 @@
       "types": "./src/archive.ts",
       "default": "./src/archive.ts"
     },
+    "./code-signature": {
+      "types": "./src/code-signature.ts",
+      "default": "./src/code-signature.ts"
+    },
+    "./code-signature-cache": {
+      "types": "./src/code-signature-cache.ts",
+      "default": "./src/code-signature-cache.ts"
+    },
     "./command": {
       "types": "./src/command.ts",
       "default": "./src/command.ts"
